First whack at new `UsersController::Action::Edit`
Removed `lib/newpoc-action-user-edit`, using new class in `#edit`
action of controller. Now to clean that up.

RSpec: 715 examples, 0 failures
RuboCop: 242 files inspected, no offenses detected
