+++ Gemfile
@@ -19,7 +19,6 @@ source 'https://rails-assets.org'
   'action-session-create',
   'action-session-destroy',
   'action-session-new',
-  'action-user-edit',
   'entity-post',
   'entity-user',
   'services-markdown_html_converter'
