--- lib/newpoc-action-user-edit/.rspec
@@ -1,2 +1,0 @@
---format documentation
---color
