--- lib/newpoc-action-user-edit/Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-# Specify your gem's dependencies in newpoc-action-user-edit.gemspec
-gemspec
