--- lib/newpoc-action-user-edit/lib/newpoc/action/user/edit/version.rb
@@ -1,11 +1,0 @@
-
-module Newpoc
-  module Action
-    module User
-      # Edit-user domain-logic setup verifies that a user is logged in.
-      class Edit
-        VERSION = '0.0.1'
-      end
-    end
-  end
-end
